Add Simple Animation in the transitions between Views, Fix struct the response.
